+++ ircbot/__init__.py
@@ -1,3 +1,4 @@
 import sys, os
 drivepath = os.getcwd() + '/googledrive'
 sys.path.append(drivepath)
+
