+++ ircbot/commands.py
