+++ ircbot/sendemail.py
