+++ ircbot/stats.py
