+++ ircbot/storage.py
