+++ ircbot/utils.py
