+++ ircbot/vote.py
